aptly-rest: Reduce the 'aptlyrestcli' example down to 'scanner'
aptlyctl can now do everything else.
